requirements.txt
# Yarnnn Agent Deployment Service Dependencies
# Generated from pyproject.toml

# Core library (from open source)
claude-agent-sdk @ git+https://github.com/Kvkthecreator/claude-agentsdk-opensource.git@v0.1.0

# Web service
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# HTTP client
httpx>=0.25.0

# Async support
aiohttp>=3.9.0
